chesstournament/requirements.txt
asgiref==3.7.2
certifi==2024.6.2
cffi
charset-normalizer==3.3.2
chess==1.10.0
coverage==6.4.3
cryptography==42.0.8
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.2
# Django==4.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-templated-mail==1.1.1
djangorestframework
djangorestframework-simplejwt
djoser==2.2.3
flake8==6.1.0
gunicorn==21.2.0
idna==3.7
mccabe==0.7.0
networkx==3.3
oauthlib==3.2.2
psycopg2-binary==2.9.7
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.8.0
python-chess==1.999
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
trf==1.1.1
typing_extensions==4.9.0
urllib3==2.2.2
whitenoise==5.2.0